--- app/src/main/java/com/example/quickmaths/BindingAdapters.kt
@@ -1,2 +1,0 @@
-package com.example.quickmaths
-
