+++ app/src/main/java/com/example/quickmaths/ui/LeaderFragment.kt
@@ -1,12 +1,12 @@
-package com.example.quickmaths.leader
+package com.example.quickmaths.ui
 
-import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.quickmaths.R
+import com.example.quickmaths.viewmodels.LeaderViewModel
 
 class LeaderFragment : Fragment() {
