+++ app/src/main/java/com/example/quickmaths/ui/LostFragment.kt
@@ -1,4 +1,4 @@
-package com.example.quickmaths.lost
+package com.example.quickmaths.ui
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -6,12 +6,10 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModel
 import androidx.navigation.fragment.findNavController
-import com.example.quickmaths.R
 import com.example.quickmaths.databinding.LostFragmentBinding
+import com.example.quickmaths.viewmodels.LostViewModel
 
 class LostFragment : Fragment() {
