+++ app/src/main/java/com/example/quickmaths/ui/PlayFragment.kt
@@ -1,4 +1,4 @@
-package com.example.quickmaths.play
+package com.example.quickmaths.ui
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -6,11 +6,10 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
-import com.example.quickmaths.R
 import com.example.quickmaths.databinding.PlayFragmentBinding
+import com.example.quickmaths.viewmodels.PlayViewModel
 
 class PlayFragment : Fragment() {
 
@@ -40,7 +39,11 @@ class PlayFragment : Fragment() {
         viewModel.getScoreAndNavigateToLost.observe(viewLifecycleOwner,
         Observer {  score ->
             val navController = findNavController()
-            navController.navigate(PlayFragmentDirections.actionPlayFragmentToLostFragment(score))
+            navController.navigate(
+                PlayFragmentDirections.actionPlayFragmentToLostFragment(
+                    score
+                )
+            )
         })
 
     }
