+++ app/src/main/java/com/example/quickmaths/ui/SettingsFragment.kt
@@ -1,14 +1,14 @@
-package com.example.quickmaths.settings
+package com.example.quickmaths.ui
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
 import android.view.*
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModel
 import androidx.navigation.findNavController
 import androidx.navigation.ui.NavigationUI
 import com.example.quickmaths.R
 import com.example.quickmaths.databinding.SettingsFragmentBinding
+import com.example.quickmaths.viewmodels.SettingsViewModel
 
 class SettingsFragment : Fragment() {
