+++ app/src/main/java/com/example/quickmaths/util/BindingAdapters.kt
@@ -1,0 +1,2 @@
+package com.example.quickmaths.util
+
