+++ app/src/main/java/com/example/quickmaths/viewmodels/LostViewModel.kt
@@ -1,13 +1,15 @@
-package com.example.quickmaths.lost
+package com.example.quickmaths.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import com.example.quickmaths.SingleLiveEvent
+import com.example.quickmaths.util.SingleLiveEvent
 
 class LostViewModel : ViewModel() {
 
-    private val onNavigateToPlayFragment = SingleLiveEvent<Void>()
-    private val onNavigateToLeaderFragment = SingleLiveEvent<Void>()
+    private val onNavigateToPlayFragment =
+        SingleLiveEvent<Void>()
+    private val onNavigateToLeaderFragment =
+        SingleLiveEvent<Void>()
 
     fun onNavigateToPlayFragment(): LiveData<Void>{
         return onNavigateToPlayFragment
