+++ app/src/main/res/layout/home_fragment.xml
@@ -2,12 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".home.HomeFragment">
+    tools:context=".ui.HomeFragment">
 
     <data>
         <variable
             name="viewModel"
-            type="com.example.quickmaths.home.HomeViewModel" />
+            type="com.example.quickmaths.viewmodels.HomeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
