+++ app/src/main/res/layout/leader_fragment.xml
@@ -2,12 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".leader.LeaderFragment">
+    tools:context=".ui.LeaderFragment">
 
     <data>
         <variable
             name="viewModel"
-            type="com.example.quickmaths.leader.LeaderViewModel" />
+            type="com.example.quickmaths.viewmodels.LeaderViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
