+++ app/src/main/res/layout/lost_fragment.xml
@@ -2,12 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".lost.LostFragment">
+    tools:context=".ui.LostFragment">
 
     <data>
         <variable
             name="viewModel"
-            type="com.example.quickmaths.lost.LostViewModel" />
+            type="com.example.quickmaths.viewmodels.LostViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
