+++ app/src/main/res/layout/play_fragment.xml
@@ -2,13 +2,13 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".play.PlayFragment">
+    tools:context=".ui.PlayFragment">
 
     <data>
 
         <variable
             name="viewModel"
-            type="com.example.quickmaths.play.PlayViewModel" />
+            type="com.example.quickmaths.viewmodels.PlayViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
