+++ app/src/main/res/layout/settings_fragment.xml
@@ -2,12 +2,12 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".settings.SettingsFragment">
+    tools:context=".ui.SettingsFragment">
 
     <data>
         <variable
             name="viewModel"
-            type="com.example.quickmaths.settings.SettingsViewModel" />
+            type="com.example.quickmaths.viewmodels.SettingsViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
